requirements.txt
transformers
python-dotenv
langchain
openai
google-api-python-client # Google API
google-search-results # SerpAPI
youtube-transcript-api
chromadb
tiktoken
faiss-cpu
wikipedia
unstructured
pdf2image
huggingface_hub
pypdf
stream-chat
streamlit_js_eval
pytube
